+++ lib/mic_stream.dart
@@ -10,9 +10,16 @@ class Microphone implements StreamController {
   StreamController<Uint8List> _controller;  // Internal implemented Controller
 
   static const _platform = const MethodChannel('mic_stream');
-  static const DEFAULT_SAMPLE_RATE = 32000;
+
+  static const DEFAULT_AUDIO_SOURCE = 1;      // Mic
+
+  static const DEFAULT_CHANNEL_CONFIG = 16;   // Mono
+
+  static const DEFAULT_AUDIO_FORMAT = 3;      // 8-bit PCM
+
+  static const DEFAULT_SAMPLE_RATE = 32000;   //in HZ
   static const MIN_SAMPLE_RATE = 1;
-  static const MAX_SAMPLE_RATE = 48000;
+  static const MAX_SAMPLE_RATE = 96000;
 
   bool _isRecording = false;
   bool _isRunning = false;
@@ -30,7 +37,9 @@ class Microphone implements StreamController {
 
   // Starts and returns an audio stream from the microphone, which is given back as Uint8List; each element has the length of .bufferSize
   // Throws an ArgumentError if Sample Rate is not between 1 and 16000
-  Future<Stream<Uint8List>> start({int sampleRate = DEFAULT_SAMPLE_RATE}) async {
+  Future<Stream<Uint8List>> start(
+      {int audioSource = DEFAULT_AUDIO_SOURCE, int sampleRate = DEFAULT_SAMPLE_RATE, int channelConfig = DEFAULT_CHANNEL_CONFIG, int audioFormat = DEFAULT_AUDIO_FORMAT}
+      ) async {
     if (!_isRecording) {
       _isRecording = true;
 
@@ -40,8 +49,20 @@ class Microphone implements StreamController {
       print("mic_stream: Set timestamp");
       _timestamp = new DateTime.now();
 
+      print("mic_stream: Set audio source");
+      await setAudioSource(audioSource);
+
       print("mic_stream: Set sample rate");
-      await _platform.invokeMethod('setSampleRate', <String, int>{'sampleRate': sampleRate});
+      await setSampleRate(sampleRate);
+
+      print("mic_stream: Set channel config");
+      await setChannelConfig(channelConfig);
+
+      print("mic_stream: Set audio format");
+      await setAudioFormat(audioFormat);
+
+      print("mic_stream: Update buffer size");
+      _bufferSize = await bufferSize;
 
       print("mic_stream: Init Audio Recorder");
       try {
@@ -51,9 +72,6 @@ class Microphone implements StreamController {
         print("mic_stream: IOERROR - Could not initialize audio recorder");
         throw(StateError);
       }
-
-      print("mic_stream: Update buffer size");
-      _bufferSize = await bufferSize;
 
       print("mic_stream: Start recording:");
       _run();
@@ -109,10 +127,16 @@ class Microphone implements StreamController {
   }
 
   // Changes the sample rate (only necessary for changing while recording - might cause unintended behaviour)
-  set sampleRate(int sampleRate) {
+  void setSampleRate(int sampleRate) {
     if (sampleRate < MIN_SAMPLE_RATE || sampleRate > MAX_SAMPLE_RATE) throw(RangeError.range(sampleRate, MIN_SAMPLE_RATE, MAX_SAMPLE_RATE));
     _platform.invokeMethod('setSampleRate', <String, int>{'sampleRate': sampleRate});
   }
+
+  void setAudioSource (int audioSource) => _platform.invokeMethod('setAudioSource', <String, int>{'audioSource': audioSource});
+
+  void setChannelConfig (int channelConfig) => _platform.invokeMethod('setChannelConfig', <String, int>{'channelConfig': channelConfig});
+
+  void setAudioFormat (int audioFormat) => _platform.invokeMethod('setAudioFormat', <String, int>{'audioFormat': audioFormat});
 
   bool get isRecording => _isRecording;